requirements.txt
# -------------------------------------------------
# Core LangChain (Modular v0.2.x)
# -------------------------------------------------
langchain==0.2.1
# Let langchain / langgraph / langchain-openai agree on langchain-core version
langchain-community==0.2.1
langchain-text-splitters==0.2.0

# -------------------------------------------------
# Vector Store
# -------------------------------------------------
chromadb==0.5.0
langchain-chroma==0.1.4

# -------------------------------------------------
# Document Loaders
# -------------------------------------------------
pypdf==4.2.0
unstructured==0.21.5
python-docx==1.1.0

# -------------------------------------------------
# OpenAI (Embeddings / Chat / Evaluation)
# -------------------------------------------------
langchain-openai==0.1.8

# -------------------------------------------------
# LangGraph Orchestration
# -------------------------------------------------
langgraph==0.2.68

# -------------------------------------------------
# FastAPI Backend + UI
# -------------------------------------------------
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
jinja2==3.1.3

# -------------------------------------------------
# Utilities
# -------------------------------------------------
python-dotenv==1.0.1

# -------------------------------------------------
# RAGAS Evaluation
# -------------------------------------------------
ragas
datasets
# tqdm is pulled in transitively by chromadb / unstructured
# Let pip choose a compatible version instead of pinning it
